Fix init order.
