+++ Engine/JocysCom.VS.AiCompanion.Engine.csproj
@@ -72,8 +72,8 @@
 		<PackageReference Include="MailKit" Version="4.8.0" />
 		<PackageReference Include="Microsoft.Bcl.AsyncInterfaces" Version="8.0.0" />
 		<PackageReference Include="Microsoft.CognitiveServices.Speech" Version="1.41.1" />
-		<PackageReference Include="Microsoft.Graph" Version="5.61.0" />
-		<PackageReference Include="Microsoft.IdentityModel.JsonWebTokens" Version="8.1.2" />
+		<PackageReference Include="Microsoft.Graph" Version="5.62.0" />
+		<PackageReference Include="Microsoft.IdentityModel.JsonWebTokens" Version="8.2.0" />
 		<PackageReference Include="Microsoft.ML" Version="3.0.1" />
 		<PackageReference Include="Microsoft.OpenApi" Version="1.6.22" />
 		<PackageReference Include="Microsoft.OpenApi.Readers" Version="1.6.22" />
@@ -85,7 +85,7 @@
 		<PackageReference Include="System.Composition" Version="8.0.0" />
 		<PackageReference Include="System.Net.Http" Version="4.3.4" />
 		<PackageReference Include="System.Runtime.Loader" Version="4.3.0" />
-		<PackageReference Include="YamlDotNet" Version="16.1.3" />
+		<PackageReference Include="YamlDotNet" Version="16.2.0" />
 	</ItemGroup>
 
 	<ItemGroup Condition=" '$(TargetFramework)' == 'net48' ">
