+++ Plugins/Core/JocysCom.VS.AiCompanion.Plugins.Core.csproj
@@ -56,24 +56,28 @@
 		  <IncludeAssets>none</IncludeAssets>
 		</PackageReference>
 		<PackageReference Include="ColorCode.HTML" Version="2.0.15" />
-		<PackageReference Include="SkiaSharp.NativeAssets.macOS" Version="2.88.8" ExcludeAssets="All" PrivateAssets="All" />
-		<PackageReference Include="SkiaSharp.NativeAssets.Win32" Version="2.88.8" GeneratePathProperty="true" ExcludeAssets="All" />
+		<PackageReference Include="SkiaSharp.NativeAssets.macOS" Version="2.88.9" ExcludeAssets="All" PrivateAssets="all">
+		  <IncludeAssets>none</IncludeAssets>
+		</PackageReference>
+		<PackageReference Include="SkiaSharp.NativeAssets.Win32" Version="2.88.9" GeneratePathProperty="true" ExcludeAssets="All" />
 		<PackageReference Include="bblanchon.PDFium.Linux" Version="130.0.6721" ExcludeAssets="All" PrivateAssets="all">
 		  <IncludeAssets>none</IncludeAssets>
 		</PackageReference>
 		<PackageReference Include="bblanchon.PDFium.Win32" Version="130.0.6721" GeneratePathProperty="true" ExcludeAssets="All" />
-		<PackageReference Include="SkiaSharp.NativeAssets.Linux.NoDependencies" Version="2.88.8" ExcludeAssets="All" PrivateAssets="All" />
+		<PackageReference Include="SkiaSharp.NativeAssets.Linux.NoDependencies" Version="2.88.9" ExcludeAssets="All" PrivateAssets="all">
+		  <IncludeAssets>none</IncludeAssets>
+		</PackageReference>
 		<PackageReference Include="BouncyCastle.Cryptography" Version="2.4.0" />
 		<PackageReference Include="DocumentFormat.OpenXml" Version="3.1.1" />
 		<PackageReference Include="HtmlAgilityPack" Version="1.11.70" />
 		<PackageReference Include="LiteDB" Version="5.0.21" />
-		<PackageReference Include="NPOI" Version="2.7.1" />
+		<PackageReference Include="NPOI" Version="2.7.2" />
 		<PackageReference Include="PdfPig" Version="0.1.9" />
 		<PackageReference Include="PDFtoImage" Version="4.1.1" />
 		<PackageReference Include="SixLabors.ImageSharp" Version="3.1.5" />
 		<PackageReference Include="System.Data.OleDb" Version="8.0.1" />
 		<PackageReference Include="System.Formats.Asn1" Version="8.0.1" />
-		<PackageReference Include="Tiktoken" Version="2.0.3" />
+		<PackageReference Include="Tiktoken" Version="2.1.1" />
 	</ItemGroup>
 
 	<ItemGroup>
