+++ Plugins/TeamsPlugin/JocysCom.VS.AiCompanion.Plugins.TeamsPlugin.csproj
@@ -30,8 +30,8 @@
 
 	<ItemGroup>
 		<PackageReference Include="Microsoft.Identity.Client" Version="4.66.1" />
-		<PackageReference Include="Microsoft.Identity.Web" Version="3.2.2" />
-		<PackageReference Include="Microsoft.Identity.Web.UI" Version="3.2.2" />
+		<PackageReference Include="Microsoft.Identity.Web" Version="3.3.1" />
+		<PackageReference Include="Microsoft.Identity.Web.UI" Version="3.3.1" />
 		<PackageReference Include="Swashbuckle.AspNetCore" Version="6.9.0" />
 	</ItemGroup>
